Split the cloudevent partial definitions to different files
This is a quality-of-life improvement that splits up the giant
`src/mock-cloud-event-partial-definitions.ts` by event type.

`src/mock-cloud-event-partial-definitions.ts` is now just `partials.ts`,
and it returns a list of partials that are all definined in different
files.

Its not the DRYest thing in thr world... but its much more readable.
